+++ plugin-dev/Source/Sentry/Private/Interface/SentryTransactionInterface.h
@@ -12,7 +12,9 @@ public:
 	virtual ~ISentryTransaction() = default;
 
 	virtual TSharedPtr<ISentrySpan> StartChild(const FString& operation, const FString& desctiption) = 0;
+	virtual TSharedPtr<ISentrySpan> StartChildWithTimestamp(const FString& operation, const FString& desctiption, int64 timestamp) = 0;
 	virtual void Finish() = 0;
+	virtual void FinishWithTimestamp(int64 timestamp) = 0;
 	virtual bool IsFinished() const = 0;
 	virtual void SetName(const FString& name) = 0;
 	virtual void SetTag(const FString& key, const FString& value) = 0;